Custom airgap (#4)
* tar archive for each component

* online serve packages from S3 redirect

* shell job to build airgap bundles

* build airgap bundles at runtime

All supported versions of Kubernetes, Docker and addons are bundled into
individual tar archives in CI and uploaded to S3. When a new installer
is created, the server will create a Job to generate an airgap bundle.
The server passes the packages required for the installer to the Job
along with a signed URL for S3 where it can upload the bundle. The Job
is a simple bash script that downloads all the packages and reassembles
into a single archive.

The online install script uses the same packages used in the airgap
bundle but downloads them at runtime. The online server redirects all
dist/ (pacakge) and bundle/ requests to S3.

Future work:
- For new named installers the bundle may exist with a different ID. It
would be more efficient to re-use bundles when available.
- After POSTing a new installer it takes a minute for the airgap bundle
to be ready.
- There's no reconciler that detects failed airgap bundles and retries
or escalates.

+++ scripts/common/addon.sh
@@ -12,14 +12,26 @@ function addon() {
     rm -rf $DIR/kustomize/$name
     mkdir -p $DIR/kustomize/$name
 
-    if [ "$AIRGAP" != "1" ] && [ -n "$KURL_URL" ]; then
-        curl -O "$KURL_URL/dist/addons/$name-$version.tar.gz"
-        mkdir -p $DIR/addons/$name/$version
-        tar xf $name-$version.tar.gz -C $DIR/addons/$name/$version
-        rm $name-$version.tar.gz
-    fi
+    addon_load $name $version
 
     . $DIR/addons/$name/$version/install.sh
 
     $name
+}
+
+function addon_load() {
+    local name=$1
+    local version=$2
+
+    if [ -z "$version" ]; then
+        return 0
+    fi
+
+    if [ "$AIRGAP" != "1" ] && [ -n "$KURL_URL" ]; then
+        curl -sSLO "$KURL_URL/dist/$name-$version.tar.gz"
+        tar xf $name-$version.tar.gz
+        rm $name-$version.tar.gz
+    fi
+
+    load_images $DIR/addons/$name/$version/images
 }
